Refactor base impl out of coresnet main and rename frames_per_clip
