+++ models/common/__init__.py
@@ -1,1 +1,2 @@
+from .base import Co3dBase  # noqa: F401
 from .res import CoResBlock, CoResStage, c2_msra_fill, init_weights  # noqa: F401
