+++ models/coresnet/scripts/profile/charades.sh
@@ -55,7 +55,7 @@ python $PROJECT/main.py \
     --logging_backend wandb \
     --num_workers 4 \
     --precision $PRECISION \
-    --frames_per_clip 64 \
+    --temporal_window_size 64 \
     --co3d_forward_mode frame \
     --profile_model \
